requirements.txt
aiohttp==3.8.5
aiosignal==1.2.0
alembic==1.11.1
anyio==3.7.0
apache-airflow==2.6.2
apache-airflow-providers-cncf-kubernetes==7.1.0
apache-airflow-providers-common-sql==1.5.2
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.4.2
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.2
apispec==5.2.2
appnope==0.1.3
appscript==1.2.2
argcomplete==3.1.1
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.12.1
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.1.3
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.0.7
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.4
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
discord.py==2.3.2
Django==4.2.3
django-picklefield==3.1
django-q==1.3.9
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
ezodf==0.3.2
fake-useragent==1.2.1
fastjsonschema==2.18.0
filelock==3.4.1
finlab==0.4.5
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.39.4
frozenlist==1.2.0
google-auth==2.21.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
ipykernel==6.23.1
ipython==8.14.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.1
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kaleido==0.2.1
kiwisolver==1.4.4
kubernetes==23.6.0
kubernetes-asyncio==24.2.3
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lml==0.1.0
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.2
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
msgpack==1.0.5
multidict==5.2.0
nbformat==5.9.2
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.9.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.5.1
plotly==5.16.1
pluggy==1.2.0
prison==0.2.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.6
Pygments==2.15.1
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysolace==0.9.30
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.1.0
redis==3.5.3
regex==2021.11.10
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.4.2
rich-argparse==1.2.0
rsa==4.9
sentry-sdk==1.14.0
setproctitle==1.3.2
shioaji==1.1.13
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.0.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
ujson==5.4.0
unicodecsv==0.14.1
urllib3==2.0.6
wcwidth==0.2.6
websocket-client==1.6.1
websockets==9.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.15.0
WTForms==3.0.1
xlrd==2.0.1
xxhash==3.2.0
yarl==1.7.2
zipp==3.15.0
